requirements_dev.txt
bumpversion==0.5.3
wheel==0.29.0
watchdog==0.8.3
flake8==3.2.1
tox==2.5.0
coverage==4.2
Sphinx==1.4.5
PyYAML==3.13
cryptography==2.3
requests-mock==1.2.0
setuptools==26.1.1
beautifulsoup4==4.6.0
